// src/types/types.ts
export interface JsonObject {
  [name: string]: any // TODO: this needs to become `unknown`
}
// ...
export interface EdgeFakeUser {
  username: string
  loginId: string
  loginKey: string
  repos: { [repo: string]: { [path: string]: JsonObject } }
  server: JsonObject
}